Redesign: contacts_book responsible for adding contacts to a db.
